feat: user.model 타입 따로 빼기
