+++ src/models/user.model.ts
@@ -1,0 +1,13 @@
+export interface ProfileData {
+  name: string;
+  sort: string;
+  bio: string;
+  git: string;
+}
+
+export interface FollowingData {
+  id: number;
+  name: string;
+  email: string;
+  follow: boolean;
+}
